remove GeneralSettings abstraction
